refactor: 번호 생성 담당 변경 (Ticket -> TicketSellingMachine)
