+++ src/test/java/step2/ResultAnalyzerTest.java
@@ -3,6 +3,7 @@ package step2;
 import org.junit.jupiter.api.Test;
 import step2.domain.ResultAnalyzer;
 import step2.domain.Ticket;
+import step2.domain.TicketSellingMachine;
 
 import java.util.Arrays;
 
@@ -12,25 +13,25 @@ class ResultAnalyzerTest {
 
     @Test
     void getMatchTicketCount() {
-        Ticket matchTicket = new Ticket();
-        int[] winningNumber = matchTicket.getNumbers();
+        Ticket ticket = TicketSellingMachine.ticketing();
+        int[] winningNumber = ticket.getNumbers();
 
         assertThat(
-                ResultAnalyzer.getMatchTicketCount(winningNumber, Arrays.asList(matchTicket), winningNumber.length)
+                ResultAnalyzer.getMatchTicketCount(winningNumber, Arrays.asList(ticket), winningNumber.length)
         ).isEqualTo(1);
 
         assertThat(
-                ResultAnalyzer.getMatchTicketCount(winningNumber, Arrays.asList(matchTicket), winningNumber.length - 1)
+                ResultAnalyzer.getMatchTicketCount(winningNumber, Arrays.asList(ticket), winningNumber.length - 1)
         ).isEqualTo(0);
     }
 
     @Test
     void getRateReturn() {
-        Ticket matchTicket = new Ticket();
-        int[] winningNumber = matchTicket.getNumbers();
+        Ticket ticket = TicketSellingMachine.ticketing();
+        int[] winningNumber = ticket.getNumbers();
 
         assertThat(
-                ResultAnalyzer.getRateReturn(winningNumber, Arrays.asList(matchTicket), matchTicket.prizeMoney(winningNumber))
+                ResultAnalyzer.getRateReturn(winningNumber, Arrays.asList(ticket), ticket.prizeMoney(winningNumber))
         ).isEqualTo(1.0);
     }
 }
