+++ src/test/java/step2/TicketSellingMachineTest.java
@@ -1,5 +1,6 @@
 package step2;
 
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import step2.domain.Ticket;
@@ -11,10 +12,18 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 class TicketSellingMachineTest {
 
+    @Test
+    void ticketing() {
+        Ticket ticket = TicketSellingMachine.ticketing();
+
+        assertThat(ticket.getNumbers().length).isEqualTo(6);
+    }
+
     @ParameterizedTest
     @CsvSource(value = {"900=0", "1000=1", "10000=10"}, delimiter = '=')
     void buy(int money, int ticketCount) {
         List<Ticket> tickets = TicketSellingMachine.buy(money);
+
         assertThat(tickets.size()).isEqualTo(ticketCount);
     }
 }
